+++ docs/changelog.md
@@ -29,6 +29,8 @@
 - Configured Github Action bot as author for the testing branch commits
 - Workflows automatically fetch all possible python versions from the `pyproject.toml` for testing, tests that only use a singular Python version now use the latest possible Python version
 - Tests now use `QudiKernel` instead of a custom remote server
+- Added install script
+- Updated install documentation
 
 ## Version 0.6.0